requirements.txt
langchain-community
langchain-core
langchain-text-splitters
langchain-ollama
langchain-huggingface
python-dotenv
requests
faiss-cpu
uuid



# absl-py
# aiohttp
# altair
# # anaconda-client==1.12.0
# anaconda-client
# anaconda-navigator
# annotated-types
# appdirs
# asgiref
# astunparse
# async-timeout
# asyncpg
# atomicwrites
# attrs
# azure-core
# azure-identity
# backoff
# backports.weakref
# bcrypt
# beautifulsoup4
# bertopic
# blinker
# blis
# branca
# brotlipy
# build
# cachetools
# Cartopy
# catalogue
# charset-normalizer
# chroma-hnswlib
# chromadb
# click-plugins
# cligj
# cloudpathlib
# clyent
# coloredlogs
# conda-repo-cli
# conda-verify
# confection
# constantly
# cssselect
# cymem
# Cython
# dataclasses-json
# datashape
# Deprecated
# dirtyjson
# distlib
# distro
# dnspython
# docopt
# editdistance
# email_validator
# en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl#sha256=ab70aeb6172cde82508f7739f35ebc9918a3d07debeed637403c8f794ba3d3dc
# en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl#sha256=6215d71a3212690e9aec49408a27e3fe6ad7cd6c715476e93d70dc784041e93e
# et-xmlfile
# fastapi
# fastapi-cli
# filelock
# fiona
# flatbuffers
# folium
# fonttools
# fredapi
# fsspec
# gast
# geopandas
# gitdb
# GitPython
# google-auth
# google-pasta
# googleapis-common-protos
# grpcio
# h11
# h2
# h5py
# hdbscan
# hpack
# htmltools
# httpcore
# httplib2
# httptools
# httpx
# huggingface-hub
# humanfriendly
# hyperframe
# importlib_resources
# inflection
# ipyleaflet
# ipython
# jsonpatch
# jsonpointer
# keras
# kubernetes
# langchain
# langchain-community
# langchain-core
# langchain-experimental
# langchain-openai
# langchain-text-splitters
# langcodes
# langgraph
# langgraph-checkpoint
# langgraph-prebuilt
# langgraph-sdk
# langsmith
# Levenshtein
# libclang
# llama-index
# llama-index-agent-openai
# llama-index-cli
# llama-index-core
# llama-index-embeddings-openai
# llama-index-graph-stores-nebula
# llama-index-indices-managed-llama-cloud
# llama-index-legacy
# llama-index-llms-azure-openai
# llama-index-llms-openai
# llama-index-multi-modal-llms-openai
# llama-index-program-openai
# llama-index-question-gen-openai
# llama-index-readers-file
# llama-index-readers-llama-parse
# llama-index-vector-stores-chroma
# llama-index-vector-stores-pinecone
# llama-index-vector-stores-postgres
# llama-parse
# llamaindex-py-client
# marshmallow
# mistune
# ml-dtypes
# mmh3
# monotonic
# msal
# msal-extensions
# munkres
# murmurhash
# mypy-extensions
# namex
# navigator-updater
# nbconvert
# nebula3-python
# nest-asyncio
# numpy
# oauthlib
# ollama
# onnxruntime
# openai
# openpyxl
# opentelemetry-api
# opentelemetry-exporter-otlp-proto-common
# opentelemetry-exporter-otlp-proto-grpc
# opentelemetry-instrumentation
# opentelemetry-instrumentation-asgi
# opentelemetry-instrumentation-fastapi
# opentelemetry-proto
# opentelemetry-sdk
# opentelemetry-semantic-conventions
# opentelemetry-util-http
# opt-einsum
# optree
# orjson
# ormsgpack
# outcome
# overrides
# packaging
# pandas-datareader
# patsy
# pdfminer
# pdfminer.six
# pdfplumber
# pep8
# pgvector
# Pillow
# pinecone-client
# pipenv
# pipreqs
# platformdirs
# ply
# portalocker
# posthog
# preshed
# protobuf
# psycopg2-binary
# pyarrow
# pyasn1-modules
# PyAudio
# pycryptodome
# pycurl
# pydantic
# pydantic_core
# pydeck
# PyDispatcher
# pyls-spyder
# Pympler
# pynndescent
# pypdf
# pypdfium2
# PyPika
# pyproj
# pyproject_hooks
# PyQt5-sip
# pyshp
# python-dotenv
# python-lsp-jsonrpc
# python-multipart
# pytz-deprecation-shim
# PyYAML
# queuelib
# rank-bm25
# rapidfuzz
# requests-oauthlib
# rich
# rsa
# safetensors
# scipy
# selenium

# sentence-transformers
# shapely
# shellingham
# shiny
# shinywidgets
# smmap
# sniffio
# sodapy